Update README.md
Wanted to add a little more clarity around the build process. When I was building this the first time, it was quite hard. And I was switching between the instructions a lot. This little customization would have helped me as it shows a link between the two set's of instructions. Also I think the build all in one with the new --experimental-targets=Xtensa configure flag is a great upgrade.